+++ c/column_fw.cc
@@ -317,9 +317,10 @@ RowIndex FwColumn<T>::join(const Column* keycol) const {
 template <typename T>
 void FwColumn<T>::fill_na_mask(int8_t* outmask, size_t row0, size_t row1) {
   const T* tdata = elements_r();
-  for (size_t i = row0; i < row1; ++i) {
-    outmask[i] = ISNA<T>(tdata[i]);
-  }
+  ri.iterate(row0, row1, 1,
+    [&](size_t i, size_t j) {
+      outmask[i] = (j == RowIndex::NA) || ISNA<T>(tdata[j]);
+    });
 }