+++ c/models/murmurhash.cc
@@ -4,6 +4,7 @@
 // https://github.com/aappleby/smhasher
 //------------------------------------------------------------------------------
 #include "models/murmurhash.h"
+#include "utils/macros.h"
 
 
 /*
@@ -33,12 +34,12 @@ uint64_t hash_murmur2(const void* key, uint64_t len, unsigned int seed)
   const uint8_t* data2 = reinterpret_cast<const uint8_t*>(data);
 
   switch (len & 7) {
-    case 7: h ^= uint64_t(data2[6]) << 48; [[clang::fallthrough]];
-    case 6: h ^= uint64_t(data2[5]) << 40; [[clang::fallthrough]];
-    case 5: h ^= uint64_t(data2[4]) << 32; [[clang::fallthrough]];
-    case 4: h ^= uint64_t(data2[3]) << 24; [[clang::fallthrough]];
-    case 3: h ^= uint64_t(data2[2]) << 16; [[clang::fallthrough]];
-    case 2: h ^= uint64_t(data2[1]) << 8;  [[clang::fallthrough]];
+    case 7: h ^= uint64_t(data2[6]) << 48; FALLTHROUGH;
+    case 6: h ^= uint64_t(data2[5]) << 40; FALLTHROUGH;
+    case 5: h ^= uint64_t(data2[4]) << 32; FALLTHROUGH;
+    case 4: h ^= uint64_t(data2[3]) << 24; FALLTHROUGH;
+    case 3: h ^= uint64_t(data2[2]) << 16; FALLTHROUGH;
+    case 2: h ^= uint64_t(data2[1]) << 8;  FALLTHROUGH;
     case 1: h ^= uint64_t(data2[0]);
             h *= m;
   };
@@ -90,22 +91,22 @@ void hash_murmur3(const void* key, const uint64_t len, unsigned int seed,
   uint64_t k2 = 0;
 
   switch (len & 15) {
-    case 15: k2 ^= (static_cast<uint64_t>(tail[14])) << 48; [[clang::fallthrough]];
-    case 14: k2 ^= (static_cast<uint64_t>(tail[13])) << 40; [[clang::fallthrough]];
-    case 13: k2 ^= (static_cast<uint64_t>(tail[12])) << 32; [[clang::fallthrough]];
-    case 12: k2 ^= (static_cast<uint64_t>(tail[11])) << 24; [[clang::fallthrough]];
-    case 11: k2 ^= (static_cast<uint64_t>(tail[10])) << 16; [[clang::fallthrough]];
-    case 10: k2 ^= (static_cast<uint64_t>(tail[ 9])) << 8; [[clang::fallthrough]];
+    case 15: k2 ^= (static_cast<uint64_t>(tail[14])) << 48; FALLTHROUGH;
+    case 14: k2 ^= (static_cast<uint64_t>(tail[13])) << 40; FALLTHROUGH;
+    case 13: k2 ^= (static_cast<uint64_t>(tail[12])) << 32; FALLTHROUGH;
+    case 12: k2 ^= (static_cast<uint64_t>(tail[11])) << 24; FALLTHROUGH;
+    case 11: k2 ^= (static_cast<uint64_t>(tail[10])) << 16; FALLTHROUGH;
+    case 10: k2 ^= (static_cast<uint64_t>(tail[ 9])) << 8; FALLTHROUGH;
     case  9: k2 ^= (static_cast<uint64_t>(tail[ 8])) << 0;
-             k2 *= c2; k2 = ROTL64(k2, 33); k2 *= c1; h2 ^= k2; [[clang::fallthrough]];
+             k2 *= c2; k2 = ROTL64(k2, 33); k2 *= c1; h2 ^= k2; FALLTHROUGH;
 
-    case  8: k1 ^= (static_cast<uint64_t>(tail[ 7])) << 56; [[clang::fallthrough]];
-    case  7: k1 ^= (static_cast<uint64_t>(tail[ 6])) << 48; [[clang::fallthrough]];
-    case  6: k1 ^= (static_cast<uint64_t>(tail[ 5])) << 40; [[clang::fallthrough]];
-    case  5: k1 ^= (static_cast<uint64_t>(tail[ 4])) << 32; [[clang::fallthrough]];
-    case  4: k1 ^= (static_cast<uint64_t>(tail[ 3])) << 24; [[clang::fallthrough]];
-    case  3: k1 ^= (static_cast<uint64_t>(tail[ 2])) << 16; [[clang::fallthrough]];
-    case  2: k1 ^= (static_cast<uint64_t>(tail[ 1])) << 8; [[clang::fallthrough]];
+    case  8: k1 ^= (static_cast<uint64_t>(tail[ 7])) << 56; FALLTHROUGH;
+    case  7: k1 ^= (static_cast<uint64_t>(tail[ 6])) << 48; FALLTHROUGH;
+    case  6: k1 ^= (static_cast<uint64_t>(tail[ 5])) << 40; FALLTHROUGH;
+    case  5: k1 ^= (static_cast<uint64_t>(tail[ 4])) << 32; FALLTHROUGH;
+    case  4: k1 ^= (static_cast<uint64_t>(tail[ 3])) << 24; FALLTHROUGH;
+    case  3: k1 ^= (static_cast<uint64_t>(tail[ 2])) << 16; FALLTHROUGH;
+    case  2: k1 ^= (static_cast<uint64_t>(tail[ 1])) << 8; FALLTHROUGH;
     case  1: k1 ^= (static_cast<uint64_t>(tail[ 0])) << 0;
              k1 *= c1; k1 = ROTL64(k1, 31); k1 *= c2; h1 ^= k1;
   };